media: venus: increase compressed frame size
When encoding 1920x1080 on the Pixel 3a (SDM670), the firmware sends
buffer requirements with an encoded frame size of 2088960, while the
driver calculates a size of 1568768. Remove the math that scales the
compressed frame size by 3/4, since the result would be too small for
the encoder to output frames to.

Signed-off-by: Richard Acayan <mailingradian@gmail.com>